+++ libmov/source/mov-ftyp.c
@@ -9,19 +9,19 @@ int mov_read_ftyp(struct mov_t* mov, const struct mov_box_t* box)
 	mov->ftyp.major_brand = mov_buffer_r32(&mov->io);
 	mov->ftyp.minor_version = mov_buffer_r32(&mov->io);
 
-	for(mov->ftyp.brands_count = 0; mov->ftyp.brands_count < N_BRAND && mov->ftyp.brands_count * 4 < box->size - 8; ++mov->ftyp.brands_count)
+	for(mov->ftyp.brands_count = 0; mov->ftyp.brands_count < N_BRAND && (uint64_t)mov->ftyp.brands_count * 4 < box->size - 8; ++mov->ftyp.brands_count)
 	{
 		mov->ftyp.compatible_brands[mov->ftyp.brands_count] = mov_buffer_r32(&mov->io);
 	}
 
-	assert(box->size == 4 * mov->ftyp.brands_count + 8);
-	mov_buffer_skip(&mov->io, box->size - 4 * mov->ftyp.brands_count - 8 ); // skip compatible_brands
+	assert(box->size == 4 * (uint64_t)mov->ftyp.brands_count + 8);
+	mov_buffer_skip(&mov->io, box->size - 4 * (uint64_t)mov->ftyp.brands_count - 8 ); // skip compatible_brands
 	return 0;
 }
 
 size_t mov_write_ftyp(const struct mov_t* mov)
 {
-	size_t size, i;
+	int size, i;
 
 	size = 8/* box */ + 8/* item */ + mov->ftyp.brands_count * 4 /* compatible brands */;
 
@@ -38,7 +38,7 @@ size_t mov_write_ftyp(const struct mov_t* mov)
 
 size_t mov_write_styp(const struct mov_t* mov)
 {
-	size_t size, i;
+	int size, i;
 
 	size = 8/* box */ + 8/* item */ + mov->ftyp.brands_count * 4 /* compatible brands */;
