+++ libmov/source/mov-hdlr.c
@@ -3,8 +3,8 @@
 #include <assert.h>
 
 // 8.4.3 Handler Reference Box (p36)
-// Box Type: ¡®hdlr¡¯ 
-// Container: Media Box (¡®mdia¡¯) or Meta Box (¡®meta¡¯) 
+// Box Type: 'hdlr'
+// Container: Media Box ('mdia') or Meta Box ('meta')
 // Mandatory: Yes 
 // Quantity: Exactly one
 int mov_read_hdlr(struct mov_t* mov, const struct mov_box_t* box)
@@ -27,7 +27,7 @@ size_t mov_write_hdlr(const struct mov_t* mov)
 {
 	const struct mov_track_t* track = mov->track;
 
-	mov_buffer_w32(&mov->io, 33 + strlen(track->handler_descr)); /* size */
+	mov_buffer_w32(&mov->io, 33 + (uint32_t)strlen(track->handler_descr)); /* size */
 	mov_buffer_write(&mov->io, "hdlr", 4);
 	mov_buffer_w32(&mov->io, 0); /* Version & flags */
 
@@ -38,6 +38,6 @@ size_t mov_write_hdlr(const struct mov_t* mov)
 	mov_buffer_w32(&mov->io, 0); /* reserved */
 	mov_buffer_w32(&mov->io, 0); /* reserved */
 
-	mov_buffer_write(&mov->io, track->handler_descr, strlen(track->handler_descr)+1); /* name */
+	mov_buffer_write(&mov->io, track->handler_descr, (uint64_t)strlen(track->handler_descr)+1); /* name */
 	return 33 + strlen(track->handler_descr);
 }
