+++ librtmp/source/rtmp-handshake.c
@@ -232,7 +232,7 @@ int rtmp_handshake_c2(uint8_t* c2, uint32_t timestamp, const uint8_t* s1, size_t
 	memmove(c2, s1, bytes);
 	be_write_uint32(c2 + 4, timestamp);
 #endif
-	return bytes;
+	return (int)bytes;
 }
 
 int rtmp_handshake_s0(uint8_t* s0, int version)
@@ -277,5 +277,5 @@ int rtmp_handshake_s2(uint8_t* s2, uint32_t timestamp, const uint8_t* c1, size_t
 	memmove(s2, c1, bytes);
 	be_write_uint32(s2 + 4, timestamp);
 #endif
-	return bytes;
+	return (int)bytes;
 }
